Move calcChi to its own module
