+++ thor/orbits/__init__.py
@@ -1,6 +1,8 @@
 from .kepler import *
 from .state import *
 from .orbits import *
+from .stumpff import *
+from .chi import *
 from .universal_propagate import *
 from .aberrations import *
 from .universal_ephemeris import *
